refactor: 닉네임 중복 검증 기능 인증으로 도메인 수정
