+++ src/docs/asciidoc/auth/auth.adoc
@@ -60,3 +60,25 @@ include::{snippets}/api-v1-auth-apple-logout/request-cookies.adoc[]
 === Response
 
 include::{snippets}/api-v1-auth-apple-logout/http-response.adoc[]
+
+== *닉네임 중복 확인* ==
+
+=== 요청
+
+=== Request
+
+include::{snippets}/api-v1-auth-check-nickname/http-request.adoc[]
+
+=== Query Parameters
+
+include::{snippets}/api-v1-auth-check-nickname/query-parameters.adoc[]
+
+=== 응답
+
+=== Response
+
+include::{snippets}/api-v1-auth-check-nickname/http-response.adoc[]
+
+=== Response Fields
+
+include::{snippets}/api-v1-auth-check-nickname/response-fields.adoc[]
