+++ src/docs/asciidoc/member/member.adoc
@@ -3,7 +3,6 @@ ifndef::snippets[]
 endif::[]
 
 = 멤버 API 문서
-
 :icons: font
 :source-highlighter: highlight.js
 :toc: left
@@ -13,32 +12,6 @@ endif::[]
 == API 목록
 
 link:../foodbowl.html[API 목록으로 돌아가기]
-
-== *닉네임 중복 확인* ==
-
-=== 요청
-
-=== Request
-
-include::{snippets}/api-v1-members-check-nickname/http-request.adoc[]
-
-=== Request Headers
-
-include::{snippets}/api-v1-members-check-nickname/request-headers.adoc[]
-
-=== Query Parameters
-
-include::{snippets}/api-v1-members-check-nickname/query-parameters.adoc[]
-
-=== 응답
-
-=== Response
-
-include::{snippets}/api-v1-members-check-nickname/http-response.adoc[]
-
-=== Response Fields
-
-include::{snippets}/api-v1-members-check-nickname/response-fields.adoc[]
 
 == *회원 탈퇴* ==
