+++ src/main/java/org/dinosaur/foodbowl/domain/member/application/MemberService.java
@@ -8,7 +8,6 @@ import org.dinosaur.foodbowl.domain.blame.repository.BlameRepository;
 import org.dinosaur.foodbowl.domain.bookmark.repository.BookmarkRepository;
 import org.dinosaur.foodbowl.domain.comment.repository.CommentRepository;
 import org.dinosaur.foodbowl.domain.follow.repository.FollowRepository;
-import org.dinosaur.foodbowl.domain.member.dto.response.NicknameDuplicateCheckResponse;
 import org.dinosaur.foodbowl.domain.member.entity.Member;
 import org.dinosaur.foodbowl.domain.member.repository.MemberRepository;
 import org.dinosaur.foodbowl.domain.member.repository.MemberRoleRepository;
@@ -36,11 +35,6 @@ public class MemberService {
     private final CommentRepository commentRepository;
     private final BookmarkRepository bookmarkRepository;
     private final BlameRepository blameRepository;
-
-    public NicknameDuplicateCheckResponse checkDuplicate(String nickname) {
-        boolean hasDuplicate = memberRepository.existsByNickname(nickname);
-        return new NicknameDuplicateCheckResponse(hasDuplicate);
-    }
 
     @Transactional
     public void withDraw(Long memberId) {
