+++ src/test/java/org/dinosaur/foodbowl/domain/member/docs/MemberControllerDocsTest.java
@@ -1,25 +1,17 @@
 package org.dinosaur.foodbowl.domain.member.docs;
 
-import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.ArgumentMatchers.anyLong;
-import static org.mockito.BDDMockito.given;
 import static org.mockito.BDDMockito.willDoNothing;
 import static org.springframework.restdocs.headers.HeaderDocumentation.headerWithName;
 import static org.springframework.restdocs.headers.HeaderDocumentation.requestHeaders;
 import static org.springframework.restdocs.mockmvc.MockMvcRestDocumentation.document;
 import static org.springframework.restdocs.mockmvc.RestDocumentationRequestBuilders.delete;
-import static org.springframework.restdocs.payload.PayloadDocumentation.fieldWithPath;
-import static org.springframework.restdocs.payload.PayloadDocumentation.responseFields;
-import static org.springframework.restdocs.request.RequestDocumentation.parameterWithName;
-import static org.springframework.restdocs.request.RequestDocumentation.queryParameters;
-import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
 import static org.springframework.test.web.servlet.result.MockMvcResultHandlers.print;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
 import org.dinosaur.foodbowl.MockApiTest;
 import org.dinosaur.foodbowl.domain.member.api.MemberController;
 import org.dinosaur.foodbowl.domain.member.application.MemberService;
-import org.dinosaur.foodbowl.domain.member.dto.response.NicknameDuplicateCheckResponse;
 import org.dinosaur.foodbowl.domain.member.entity.Role.RoleType;
 import org.dinosaur.foodbowl.global.config.security.jwt.JwtTokenProvider;
 import org.junit.jupiter.api.DisplayName;
@@ -28,9 +20,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
 import org.springframework.boot.test.mock.mockito.MockBean;
 import org.springframework.http.HttpHeaders;
-import org.springframework.http.MediaType;
 import org.springframework.restdocs.headers.HeaderDescriptor;
-import org.springframework.restdocs.payload.JsonFieldType;
 
 @WebMvcTest(controllers = MemberController.class)
 public class MemberControllerDocsTest extends MockApiTest {
@@ -40,30 +30,6 @@ public class MemberControllerDocsTest extends MockApiTest {
 
     @Autowired
     private JwtTokenProvider jwtTokenProvider;
-
-    @Test
-    @DisplayName("닉네임 중복 검증을 문서화한다.")
-    void appleLogin() throws Exception {
-        String token = jwtTokenProvider.createAccessToken(1L, RoleType.ROLE_회원);
-        given(memberService.checkDuplicate(any())).willReturn(new NicknameDuplicateCheckResponse(false));
-
-        mockMvc.perform(get("/api/v1/members/check-nickname")
-                        .header("Authorization", "Bearer " + token)
-                        .queryParam("nickname", "gray")
-                        .contentType(MediaType.APPLICATION_JSON_VALUE))
-                .andDo(print())
-                .andExpect(status().isOk())
-                .andDo(document("api-v1-members-check-nickname",
-                        requestHeaders(
-                                headerWithName(HttpHeaders.AUTHORIZATION).description("서버에서 발급한 엑세스 토큰")
-                        ),
-                        queryParameters(
-                                parameterWithName("nickname").description("닉네임")
-                        ),
-                        responseFields(
-                                fieldWithPath("hasDuplicate").type(JsonFieldType.BOOLEAN).description("닉네임 중복 여부")
-                        )));
-    }
 
     @Test
     @DisplayName("회원 탈퇴를 문서화한다.")
